fix(vital-navigation): For a non-Parent menu with submenu items -- breadcrumb link is wrong
